+++ src/boolean/equal.ts
@@ -1,12 +1,13 @@
 import {curry} from '../function/curry';
+import {BinaryPred} from "../types";
 
 export const
 
     /**
      * Equality operator.
      * @function module:boolean.equal
-     * @param a {*}
-     * @param b {*}
+     * @param [a=undefined] {any}
+     * @param [b=undefined] {any}
      * @returns {boolean}
      */
-    equal = curry((a, b) => a === b);
+    equal: BinaryPred<any> = curry((a, b) => a === b) as BinaryPred<any>;