+++ tests/function/test-curry.ts
@@ -1,38 +1,43 @@
-import {curry, curry2, curry3, curry4, curry5, curryN} from '../../src/function/curry';
+import {Curry, curry, Curry1, Curry2, curry2, Curry3, curry3, curry4, Curry5, curry5, curryN} from '../../src/function';
 import {alphabetArray} from '../helpers';
+import {Binary, Unary} from "../../src/types";
 
 describe('#curryN', () => {
+    // Some funcs to use in tests
+    const recursiveBinOp = (op: Binary<number>, start: number): Binary<number> =>
+            (...args: number[]): number => args.reduce((a, b) => {
+                return op(a, b);
+            }, start),
+        multiplyRecursive = recursiveBinOp((a, b) => a * b, 1),
+        addRecursive = recursiveBinOp((a, b) => a + b, 0);
 
-    // Set curry here to use below
-    let multiplyRecursive = (...args) => args.reduce((agg, num) => num * agg, 1),
-        addRecursive = (...args) => args.reduce((agg, num) => num + agg, 0);
-
-    it ('should be of type function.', () => {
+    it('should be of type function.', () => {
         expect(curryN).toBeInstanceOf(Function);
     });
 
-    it ('should throw an error when not receiving an argument at param `0`.', () => {
+    it('should throw an error when not receiving any argument at param `0`.', () => {
         expect(curryN).toThrow();
     });
 
-    it ('should pass in any values passed in after the arity when executing the curried function', () => {
-        let add3Nums = curryN(3, addRecursive);
+    it('should pass in any values passed in after the arity when executing the curried function', () => {
+        const add3Nums = curryN(3, addRecursive),
+            addNums: Curry5<number> = add3Nums as Curry5<number>;
 
         // Curry add to add 3 numbers
-        expect(add3Nums()(1, 2, 3)) .toEqual(6);
-        expect(add3Nums(1)(2, 3))   .toEqual(6);
-        expect(add3Nums(1, 2)(3))  .toEqual(6);
-        expect(add3Nums(1, 2, 3))   .toEqual(6);
+        expect((add3Nums() as Curry3<number>)(1, 2, 3)).toEqual(6);
+        expect((add3Nums(1) as Curry2<number>)(2, 3)).toEqual(6);
+        expect((add3Nums(1, 2) as Curry1<number>)(3)).toEqual(6);
+        expect(add3Nums(1, 2, 3)).toEqual(6);
 
         // Curry `add` to add any numbers passed required arity
-        expect(add3Nums()(1, 2, 3, 5, 6))   .toEqual(17);
-        expect(add3Nums(1)(2, 3, 5, 6))     .toEqual(17);
-        expect(add3Nums(1, 2)(3, 5, 6))     .toEqual(17);
-        expect(add3Nums(1, 2, 3, 5, 6))     .toEqual(17);
+        expect((addNums() as Curry3<number>)(1, 2, 3, 4, 5)).toEqual(15);
+        expect((addNums(1) as Curry2<number>)(2, 3, 4, 5)).toEqual(15);
+        expect((addNums(1, 2) as Curry1<number>)(3, 4, 5)).toEqual(15);
+        expect(addNums(1, 2, 3, 4, 5)).toEqual(15);
     });
 
-    it ('should return a function that will not execute until the passed in "executeArity" is met.', () => {
-        let multiply5Nums = curryN(5, multiplyRecursive),
+    it('should return a function that will not execute until the passed in "executeArity" is met.', () => {
+        const multiply5Nums = curryN(5, multiplyRecursive) as Curry5<number>,
             multiplyExpectedResult = Math.pow(5, 5),
             argsToTest = [
                 [5, 5, 5, 5, 5],
@@ -41,18 +46,20 @@ describe('#curryN', () => {
                 [5, 5],
                 [5]
             ],
-            partiallyAppliedResults = [
-                multiply5Nums(),
-                multiply5Nums(5),
-                multiply5Nums(5, 5),
-                multiply5Nums(5, 5, 5),
-                multiply5Nums(5, 5, 5, 5)
+            partiallyAppliedResults: Array<Curry<number>> = [
+                multiply5Nums() as Curry<number>,
+                multiply5Nums(5) as Curry<number>,
+                multiply5Nums(5, 5) as Curry<number>,
+                multiply5Nums(5, 5, 5) as Curry<number>,
+                multiply5Nums(5, 5, 5, 5) as Curry<number>
             ];
 
         // Curry multiply and pass args in non-linear order
         argsToTest.forEach(function (args, index) {
             expect(partiallyAppliedResults[index]).toBeInstanceOf(Function);
-            expect(partiallyAppliedResults[index].apply(null, args)).toEqual(multiplyExpectedResult);
+            expect(
+                (partiallyAppliedResults[index])(...args)
+            ).toEqual(multiplyExpectedResult);
         });
 
     });
@@ -61,23 +68,26 @@ describe('#curryN', () => {
 
 describe('#curry', () => {
 
-    it ('should be of type function.', () => {
+    it('should be of type function.', () => {
         expect(curry).toBeInstanceOf(Function);
     });
 
-    it ('should return a function when receiving a function.', () => {
+    it('should return a function when receiving a function.', () => {
         expect(curry(() => undefined)).toBeInstanceOf(Function);
-        expect(curry(() => {})).toBeInstanceOf(Function);
+        expect(curry(() => {
+        })).toBeInstanceOf(Function);
     });
 
-    it ('should throw an error when receiving anything other than a function (for first param)', () => {
+    it('should throw an error when receiving anything other than a function (for first param)', () => {
         [99, false, true, null, undefined, [], {}].forEach(x => {
+            // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+            // @ts-ignore
             expect(() => curry(x)).toThrow();
         });
     });
 
-    it ('should return a curried function.', () => {
-        let min8 = curry(Math.min, 8),
+    it('should return a curried function.', () => {
+        const min8 = curry(Math.min, 8),
             max5 = curry(Math.max, 5),
             pow2 = curry(Math.pow, 2);
 
@@ -98,16 +108,18 @@ describe('#curry', () => {
         expect(pow2(4)).toEqual(16);
     });
 
-    it ('should be able to correctly curry functions of different arity as long as their arity is met.', () => {
-        let min = curry2(Math.min),
-            max = curry2(Math.max),
-            pow = curry2(Math.pow),
-            min8 = curry(Math.min, 8),
-            max5 = curry(Math.max, 5),
-            pow2 = curry(Math.pow, 2),
-            isValidTangentLen = curry((a, b, cSqrd) => pow(a, 2) + pow(b, 2) === cSqrd, 2, 2),
-            random = curry((start, end) => Math.round(Math.random() * end + start), 0),
-            expectedFor = (num) => min(8, max(5, pow(2, num)));
+    it('should be able to correctly curry functions of different arity as long as their arity is met.', () => {
+        const min = curry2(Math.min) as Curry2<number>,
+            max = curry2(Math.max) as Curry2<number>,
+            pow = curry2(Math.pow) as Curry2<number>,
+            min8 = curry(Math.min, 8) as Curry1<number>,
+            max5 = curry(Math.max, 5) as Curry1<number>,
+            pow2 = curry(Math.pow, 2) as Curry1<number>,
+            isValidTangentLen = curry((a, b, cSqrd) => <number>pow(a, 2) + <number>pow(b, 2) === cSqrd, 2, 2),
+            random = curry((start, end) => Math.round(Math.random() * (end - start) + start), 0) as Curry1<number>,
+            expectedFor: Unary<number> = num => <number>min(8, <number>max(5, <number>pow(2, num))),
+            op: Unary<number> = x2 => [min8, max5, pow2].reduceRight((agg, f) => f(agg), x2) as number
+        ;
 
         // Expect functions returned for `curry` calls
         expect(isValidTangentLen).toBeInstanceOf(Function);
@@ -122,20 +134,19 @@ describe('#curry', () => {
         expect(isValidTangentLen(21)).toEqual(false);
 
         // Expect `curry`ed functions to work as expected
-        [8,5,3,2,1,0, random(89), random(55), random(34)].forEach(function (num) {
-            let composed = x => [min8, max5, pow2].reduceRight((agg, f) => f(agg), x);
-            expect(composed(num)).toEqual(expectedFor(num));
-        });
+        [8, 5, 3, 2, 1, 0, random(89), random(55), random(34)].forEach(x =>
+            expect(op(x as number)).toEqual(expectedFor(x as number))
+        );
     });
 
 });
 
 describe('#curry2, #curry3, #curry4, #curry5', () => {
-    it ('should returned a curried function which curries 2 parameters', () => {
+    it('should returned a curried function which curries 2 parameters', () => {
         const
             min = curry2(Math.min),
             max = curry2(Math.max),
-            onlyEvens = (...args) => args.filter(x => x % 2 === 0),
+            onlyEvens = (...args): number[] => args.filter(x => x % 2 === 0),
             onlyEvens3 = curry3(onlyEvens),
             onlyEvens4 = curry4(onlyEvens),
             onlyEvens5 = curry5(onlyEvens),
@@ -152,23 +163,22 @@ describe('#curry2, #curry3, #curry4, #curry5', () => {
         expect(evenSomeNums.every(x => x % 2 === 0)).toEqual(true);
 
         // Tests table
-        //  [fn, args, expected, expectedArityAfterFirstFnParam]
+        //  [fn, args, expected]
         [
-            [min, [0, 1], 0, 1],
-            [max, [0, 1], 1, 1],
-            [max, [0, 1, 3, 5, 3, 1], 5, 1],
-            [min, [0, 1, 3, 2, 1], 0, 1],
-            [onlyEvens3, someNums, evenSomeNums, 2],
-            [onlyEvens4, someNums, evenSomeNums, 3],
-            [onlyEvens5, someNums, evenSomeNums, 4]
+            [min, [0, 1], 0],
+            [max, [0, 1], 1],
+            [max, [0, 1, 3, 5, 3, 1], 5],
+            [min, [0, 1, 3, 2, 1], 0],
+            [onlyEvens3, someNums, evenSomeNums],
+            [onlyEvens4, someNums, evenSomeNums],
+            [onlyEvens5, someNums, evenSomeNums]
         ]
-            .forEach(([fn, args, expected, expectedArityAfterFirstParam]) => {
+            .forEach(([fn, args, expected]) => {
                 expect(fn).toBeInstanceOf(Function);
                 const
                     newArgs = args.slice(0),
                     newFn = fn(newArgs.shift());
                 expect(newFn).toBeInstanceOf(Function);
-                expect(newFn.length).toEqual(expectedArityAfterFirstParam);
                 expect(newFn(...newArgs)).toEqual(expected);
             });
     });